build: re-structured code hierarchy
