+++ ospf/model/ospf-app.h
@@ -28,16 +28,16 @@
 #include "ns3/traced-callback.h"
 #include "ns3/ipv4-static-routing-helper.h"
 #include "ns3/random-variable-stream.h"
-#include "ospf-header.h"
-#include "lsa-header.h"
-#include "router-lsa.h"
-#include "as-external-lsa.h"
-#include "area-lsa.h"
-#include "summary-lsa.h"
-#include "ospf-dbd.h"
-#include "ospf-hello.h"
-#include "ls-ack.h"
-#include "ls-update.h"
+#include "ns3/ospf-header.h"
+#include "ns3/ospf-dbd.h"
+#include "ns3/ospf-hello.h"
+#include "ns3/ls-ack.h"
+#include "ns3/ls-update.h"
+#include "ns3/lsa-header.h"
+#include "ns3/router-lsa.h"
+#include "ns3/as-external-lsa.h"
+#include "ns3/area-lsa.h"
+#include "ns3/summary-lsa.h"
 #include "ospf-interface.h"
 #include "unordered_map"
 #include "queue"
