+++ ospf/model/ospf-neighbor.h
@@ -36,10 +36,10 @@
 #include "ns3/packet.h"
 #include "ns3/uinteger.h"
 #include "ns3/header.h"
-#include "lsa.h"
-#include "lsa-header.h"
-#include "ospf-dbd.h"
-#include "ls-request.h"
+#include "ns3/lsa.h"
+#include "ns3/lsa-header.h"
+#include "ns3/ospf-dbd.h"
+#include "ns3/ls-request.h"
 #include "queue"
 #include "algorithm"
