+++ ospf/model/packets/ls-update.h
@@ -24,9 +24,9 @@
 #include "ns3/object.h"
 #include "ns3/header.h"
 #include "ns3/ipv4-address.h"
-#include "ospf-interface.h"
-#include "lsa-header.h"
-#include "lsa.h"
+#include "ns3/ospf-interface.h"
+#include "ns3/lsa-header.h"
+#include "ns3/lsa.h"
 
 namespace ns3 {
 /**
