+++ ospf/wscript
@@ -6,18 +6,18 @@ def build(bld):
         'model/ospf-app.cc',
         'model/ospf-interface.cc',
         'model/ospf-neighbor.cc',
-        'model/ospf-header.cc',
-        'model/lsa-header.cc',
-        'model/lsa.cc',
-        'model/ospf-hello.cc',
-        'model/router-lsa.cc',
-        'model/as-external-lsa.cc',
-        'model/area-lsa.cc',
-        'model/summary-lsa.cc',
-        'model/ls-ack.cc',
-        'model/ls-request.cc',
-        'model/ls-update.cc',
-        'model/ospf-dbd.cc',
+        'model/packets/ospf-header.cc',
+        'model/packets/ospf-hello.cc',
+        'model/packets/ls-ack.cc',
+        'model/packets/ls-request.cc',
+        'model/packets/ls-update.cc',
+        'model/packets/ospf-dbd.cc',
+        'model/lsa/lsa-header.cc',
+        'model/lsa/lsa.cc',
+        'model/lsa/router-lsa.cc',
+        'model/lsa/as-external-lsa.cc',
+        'model/lsa/area-lsa.cc',
+        'model/lsa/summary-lsa.cc',
         'helper/ospf-app-helper.cc',
         ]
 
@@ -35,18 +35,18 @@ def build(bld):
         'model/ospf-app.h',
         'model/ospf-interface.h',
         'model/ospf-neighbor.h',
-        'model/ospf-header.h',
-        'model/lsa-header.h',
-        'model/lsa.h',
-        'model/router-lsa.h',
-        'model/as-external-lsa.h',
-        'model/area-lsa.h',
-        'model/summary-lsa.h',
-        'model/ospf-hello.h',
-        'model/ls-ack.h',
-        'model/ls-request.h',
-        'model/ls-update.h',
-        'model/ospf-dbd.h',
+        'model/packets/ospf-header.h',
+        'model/packets/ospf-hello.h',
+        'model/packets/ls-ack.h',
+        'model/packets/ls-request.h',
+        'model/packets/ls-update.h',
+        'model/packets/ospf-dbd.h',
+        'model/lsa/lsa-header.h',
+        'model/lsa/lsa.h',
+        'model/lsa/router-lsa.h',
+        'model/lsa/as-external-lsa.h',
+        'model/lsa/area-lsa.h',
+        'model/lsa/summary-lsa.h',
         'helper/ospf-app-helper.h',
         'helper/ospf-packet-helper.h',
         'helper/ospf-runtime-helper.h',
